refactor: merge @sockethub/crypto into @sockethub/util
Fold the standalone @sockethub/crypto package into @sockethub/util as a
new `@sockethub/util/crypto` subpath export, removing a package from the
build/version/publish surface.

- Move crypto source + tests to packages/util/src/crypto/
- Make encrypt/decrypt schema-agnostic (unknown) so util need not depend
  on @sockethub/schemas
- Rewire data-layer and server consumers to @sockethub/util/crypto
- Update dep graph, tsconfig paths/references, build-types script, READMEs
- Delete packages/crypto

Also tighten error handling in platform-xmpp: replace two unsafe
`done(err as Error)` casts with the shared `toError()` helper from
@sockethub/util/error.
