--- packages/crypto/package.json
@@ -1,57 +1,0 @@
-{
-  "name": "@sockethub/crypto",
-  "description": "Crypto functions and helpers for Sockethub",
-  "version": "1.0.0-alpha.12",
-  "type": "module",
-  "private": false,
-  "author": "Nick Jennings <nick@silverbucket.net>",
-  "license": "MIT",
-  "publishConfig": {
-    "access": "public"
-  },
-  "main": "dist/index.js",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "default": "./dist/index.js"
-    }
-  },
-  "files": [
-    "dist/"
-  ],
-  "engines": {
-    "node": ">=20"
-  },
-  "keywords": [
-    "sockethub",
-    "messaging",
-    "crypto",
-    "encryption",
-    "cryptography"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/sockethub/sockethub.git",
-    "directory": "packages/crypto"
-  },
-  "homepage": "https://github.com/sockethub/sockethub/tree/master/packages/crypto",
-  "scripts": {
-    "build": "bun build src/index.ts --outdir dist --target node --format esm --sourcemap=external",
-    "clean": "rm -rf dist",
-    "clean:deps": "rm -rf node_modules"
-  },
-  "dependencies": {
-    "@sockethub/schemas": "^3.0.0-alpha.12",
-    "object-hash": "^3.0.0",
-    "secure-store-redis": "^4.1.0"
-  },
-  "devDependencies": {
-    "@types/bun": "latest",
-    "@types/object-hash": "^3.0.6",
-    "@types/sinon": "^17.0.4",
-    "sinon": "^17.0.2"
-  },
-  "gitHead": "cc495a5fcf7ce19a8f328a34bbc8cd1439bb67ce",
-  "types": "./dist/index.d.ts"
-}
