+++ packages/server/package.json
@@ -52,7 +52,6 @@
   "dependencies": {
     "@sentry/node": "^10.36.0",
     "@sockethub/client": "^5.0.0-alpha.12",
-    "@sockethub/crypto": "^1.0.0-alpha.12",
     "@sockethub/data-layer": "^1.0.0-alpha.12",
     "@sockethub/logger": "^1.0.0-alpha.12",
     "@sockethub/schemas": "^3.0.0-alpha.12",
